+++ CHANGELOG.md
@@ -14,11 +14,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `sqlitex.Pool` has a new method `Take`
   which returns an `error` along with a `Conn`
   ([#83](https://github.com/zombiezen/go-sqlite/issues/83)).
+- `sqlite.ErrorOffset` is a new function
+  that returns the SQL byte offset that an error references.
+
+### Changed
+
+- `sqlite.Conn.Prep`, `sqlite.Conn.Prepare`, and `sqlite.Conn.PrepareTransient`
+  now include position information in error messages if available.
+- Many error messages around statement execution changed their format
+  for better readability.
+  Error messages are not stable API and should not be depended on.
 
 ### Deprecated
 
 - The `sqlitex.Pool.Get` method has been deprecated
   in favor of the new `Take` method.
+
+### Fixed
+
+- Error messages no longer duplicate information from their error code
+  (reported in [#84](https://github.com/zombiezen/go-sqlite/issues/84)).
 
 ## [1.1.2][] - 2024-02-14